% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pattern-geometry-crosshatch.R
\name{grid.pattern_crosshatch}
\alias{grid.pattern_crosshatch}
\title{Crosshatch patterned grobs}
\usage{
grid.pattern_crosshatch(
  x = c(0, 0.5, 1, 0.5),
  y = c(0.5, 1, 0.5, 0),
  id = 1L,
  ...,
  colour = gp$col \%||\% "grey20",
  fill = gp$fill \%||\% "grey80",
  angle = 30,
  density = 0.2,
  spacing = 0.05,
  xoffset = 0,
  yoffset = 0,
  alpha = gp$alpha \%||\% 1,
  linetype = gp$lty \%||\% 1,
  size = gp$lwd \%||\% 1,
  default.units = "npc",
  name = NULL,
  gp = gpar(),
  draw = TRUE,
  vp = NULL
)
}
\arguments{
\item{x}{A numeric vector or unit object specifying x-locations of the pattern boundary}

\item{y}{A numeric vector or unit object specifying y-locations of the pattern boundary}

\item{id}{A numeric vector used to separate locations in x, y into multiple boundaries.
All locations within the same \code{id} belong to the same boundary.}

\item{...}{Currently ignored}

\item{colour}{Stroke colour}

\item{fill}{Fill colour}

\item{angle}{Rotation angle in degrees}

\item{density}{Approx. fraction of area the pattern fills (between 0 and 1)}

\item{spacing}{Spacing between repetitions of pattern (between 0 and 1)}

\item{xoffset}{Shift pattern along x axis (between 0 and 1)}

\item{yoffset}{Shift pattern along y axis (between 0 and 1)}

\item{alpha}{Alpha (between 0 and 1)}

\item{linetype}{Stroke linetype}

\item{size}{Stroke linewidth}

\item{default.units}{A string indicating the default units to use if \code{x} or \code{y}
are only given as numeric vectors.}

\item{name}{ A character identifier. }

\item{gp}{An object of class \code{"gpar"}, typically the output
    from a call to the function \code{\link[grid]{gpar}}.  This is basically
    a list of graphical parameter settings.}

\item{draw}{A logical value indicating whether graphics output
    should be produced.}

\item{vp}{A Grid viewport object (or NULL).}
}
\value{
A grid grob object invisibly.  If \code{draw} is \code{TRUE} then also draws to the graphic device as a side effect.
}
\description{
\code{grid.pattern_crosshatch()} draws a crosshatch pattern onto the graphic device.
}
\examples{
  if (require("grid")) {
    grid.newpage()
    grid.pattern_crosshatch(colour = "green", fill = "blue", density = 0.5)
    grid.newpage()
    grid.pattern_crosshatch(density = 0.3, gp = gpar(col = "blue", fill = "yellow"))
  }
}
\seealso{
The \code{ggpattern} documentation: \url{https://coolbutuseless.github.io/package/ggpattern/articles/pattern-crosshatch.html}


}
