% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pattern-geometry-weave.R
\name{grid.pattern_weave}
\alias{grid.pattern_weave}
\title{Weave patterned grobs}
\usage{
grid.pattern_weave(
  x = c(0, 0, 1, 1),
  y = c(1, 0, 0, 1),
  id = 1L,
  ...,
  colour = gp$col \%||\% "grey20",
  fill = gp$fill \%||\% "grey80",
  fill2 = fill,
  angle = 30,
  density = 0.2,
  spacing = 0.05,
  xoffset = 0,
  yoffset = 0,
  alpha = gp$alpha \%||\% NA_real_,
  linetype = gp$lty \%||\% 1,
  linewidth = size \%||\% gp$lwd \%||\% 1,
  size = NULL,
  grid = "square",
  type = "plain",
  subtype = NA,
  default.units = "npc",
  name = NULL,
  gp = gpar(),
  draw = TRUE,
  vp = NULL
)
}
\arguments{
\item{x}{A numeric vector or unit object specifying x-locations of the pattern boundary.}

\item{y}{A numeric vector or unit object specifying y-locations of the pattern boundary.}

\item{id}{A numeric vector used to separate locations in x, y into multiple boundaries.
All locations within the same \code{id} belong to the same boundary.}

\item{...}{Currently ignored}

\item{colour}{Stroke colour}

\item{fill}{The fill colour for the horizontal "weft" lines.}

\item{fill2}{The fill colour for the vertical "warp" lines.}

\item{angle}{Rotation angle in degrees}

\item{density}{Approx. fraction of area the pattern fills.}

\item{spacing}{Spacing between repetitions of pattern ('snpc' units between 0 and 1).}

\item{xoffset}{Shift pattern along x axis ('snpc' units between 0 and 1).}

\item{yoffset}{Shift pattern along y axis ('snpc' units between 0 and 1).}

\item{alpha}{Alpha (between 0 and 1) or \code{NA} (default, preserves colors' alpha value).}

\item{linetype}{Stroke linetype}

\item{linewidth}{Stroke linewidth}

\item{size}{For backwards compatibility can be used to set \code{linewidth}}

\item{grid}{Adjusts placement and density of certain graphical elements.
\code{"square"} (default) is a square grid.
\code{"hex"} is a hexagonal grid suitable for hexagonal and triangular tiling.
\code{"hex_circle"} is a hexagonal grid suitable for circle packing.
\code{"elongated_triangle"} is a grid used for the "elongated triangle" tiling.}

\item{type}{The weave type.  See \code{\link[=pattern_weave]{pattern_weave()}} for more details.}

\item{subtype}{The weave subtype.  See \code{\link[=pattern_weave]{pattern_weave()}} for more details.}

\item{default.units}{A string indicating the default units to use if \code{x} or \code{y}
are only given as numeric vectors.}

\item{name}{ A character identifier. }

\item{gp}{An object of class \code{"gpar"}, typically the output
    from a call to the function \code{\link[grid]{gpar}}.  This is basically
    a list of graphical parameter settings.}

\item{draw}{A logical value indicating whether graphics output
    should be produced.}

\item{vp}{A Grid viewport object (or NULL).}
}
\value{
A grid grob object invisibly.  If \code{draw} is \code{TRUE} then also draws to the graphic device as a side effect.
}
\description{
\code{grid.pattern_weave()} draws a weave pattern onto the graphic device.
}
\examples{
  if (require("grid")) {
    x_hex <- 0.5 + 0.5 * cos(seq(2 * pi / 4, by = 2 * pi / 6, length.out = 6))
    y_hex <- 0.5 + 0.5 * sin(seq(2 * pi / 4, by = 2 * pi / 6, length.out = 6))
    gp <- gpar(colour = "black", fill = "lightblue", lwd=0.5)

    # Plain weave (default weave)
    grid.pattern_weave(x_hex, y_hex, fill2 = "yellow",
                       gp = gp, spacing = 0.1, density = 0.3)

    # Irregular matt weave
    grid.newpage()
    grid.pattern_weave(x_hex, y_hex,  type = "matt_irregular",
                       fill2 = "yellow", gp = gp, spacing = 0.1, density = 0.3)

    # Twill weave
    grid.newpage()
    grid.pattern_weave(x_hex, y_hex, type = "twill",
                       fill2 = "yellow", gp = gp, spacing = 0.1, density = 0.3)

    # Zig-zag twill
    grid.newpage()
    grid.pattern_weave(x_hex, y_hex, type = "twill_zigzag",
                       fill2 = "yellow", gp = gp, spacing = 0.05, density = 0.7)

    # Herringbone twill with density 1
    grid.newpage()
    gp$col <- NA
    grid.pattern_weave(x_hex, y_hex, type = "twill_herringbone",
                       fill2 = "yellow", gp = gp, spacing = 0.05, density = 1.0)

  }
}
\seealso{
\code{\link[=pattern_weave]{pattern_weave()}}
}
