% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-params.R
\name{guess_has_R4.1_features}
\alias{guess_has_R4.1_features}
\title{Guess whether "active" graphics device supports
the grid graphics features introduced in R v4.1.}
\usage{
guess_has_R4.1_features()
}
\value{
\code{TRUE} or \code{FALSE}
}
\description{
\code{guess_has_R_4.1_features()} guesses whether "active" graphics device supports
the grid graphics features introduced in R v4.1.  If it guesses it does
it returns \code{TRUE} else \code{FALSE}.
}
\examples{
  # If R version (weakly) greater than 4.1 should be TRUE
  pdf(tempfile(fileext = ".pdf"))
  print(guess_has_R4.1_features())
  invisible(dev.off())

  # Should be FALSE
  postscript(tempfile(fileext = ".ps"))
  print(guess_has_R4.1_features())
  invisible(dev.off())

}
\seealso{
\url{https://www.stat.auckland.ac.nz/~paul/Reports/GraphicsEngine/definitions/definitions.html} for more info about the new grid graphics
features introduced in R v4.1.
}
