% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_col.R
\name{mean_col}
\alias{mean_col}
\title{Compute average color}
\usage{
mean_col(...)
}
\arguments{
\item{...}{Colors to average}
}
\description{
\code{mean_col()} computes an average color.
}
\details{
We currently compute an average color
by using the quadratic mean of the colors' RGBA values.
}
\examples{
 mean_col("black", "white")
 mean_col(c("black", "white"))
 mean_col("red", "blue")
}
