% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pattern_square.R
\docType{data}
\name{pattern_square}
\alias{pattern_square}
\alias{names_square}
\title{Square pattern matrix}
\format{
An object of class \code{character} of length 2.
}
\usage{
pattern_square(type = "diagonal", subtype = NULL, nrow = 5L, ncol = 5L)

names_square
}
\arguments{
\item{type}{Either "diagonal" or "diagonal_skew" or any of the weave patterns
documented in \code{\link[=pattern_weave]{pattern_weave()}}.}

\item{subtype}{For the "diagonal" and "diagonal_skew" patterns
an integer indicating number of colors (or other graphical elements).
See \code{\link[=pattern_weave]{pattern_weave()}} for supported \code{subtype} arguments for the various weave patterns.}

\item{nrow}{Number of rows (height).}

\item{ncol}{Number of columns (width).}
}
\value{
A matrix of integer values indicating where the each color
(or other graphical element) should be drawn on a rectangular grid.
Indices \verb{[1,1]} of the matrix corresponds to the bottom-left of the grid
while indices \verb{[1,ncol]} corresponds to the bottom-right of the grid.
This matrix has a "pattern_square" subclass which supports a special \code{print()} method.
}
\description{
\code{pattern_square()} returns an integer matrix indicating where each
color (or other graphical element) should be drawn on a rectangular grid
for a specified square pattern type and subtype.
\code{names_square} lists the currently supported square \code{type}s (excluding those in \code{names_weave}).
}
\details{
"diagonal" and "diagonal_skew" simply cycle through the colors
both horizontally and vertically.
If two colors are requested this provides the standard two-color checkerboard pattern.
If there are more than three colors than "diagonal" will have colored diagonals
going from top left to bottom right while "diagonal_skew" will have them
going form bottom left to top right.

The weave patterns simply convert the logical matrix returned by \code{\link[=pattern_weave]{pattern_weave()}} into an
integer matrix by having \code{TRUE} set to \code{1L} and \code{FALSE} set to \code{2L}.
Hence the various weave patterns only support (up to) two-color patterns.
}
\examples{
 # supported square names
 print(names_square)

 # (main) diagonal has colors going from top left to bottom right
 diagonal <- pattern_square("diagonal", 4L, nrow = 7L, ncol = 9L)
 print(diagonal)

 # skew diagonal has colors going from bottom left to top right
 skew <- pattern_square("diagonal_skew", 4L, nrow = 7L, ncol = 9L)
 print(skew)

 # also supports the various 'weave' patterns
 zigzag <- pattern_square("twill_zigzag", nrow = 15L, ncol = 9L)
 print(zigzag)

}
\seealso{
\code{\link[=grid.pattern_regular_polygon]{grid.pattern_regular_polygon()}} for drawing to a graphics device
polygons in multiple color/size/shape patterns.
\code{\link[=pattern_weave]{pattern_weave()}} for more information on "weave" patterns.
}
\keyword{datasets}
