% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-ggplot2.R
\name{update_alpha}
\alias{update_alpha}
\title{Update fill colour and/or pattern transparency}
\usage{
update_alpha(fill, alpha)
}
\arguments{
\item{fill}{A fill colour given as a \code{character} or \code{integer} vector, or as a
(list of) \verb{<GridPattern>} object(s) and/or colour(s).}

\item{alpha}{A transparency value between 0 (transparent) and 1 (opaque),
parallel to \code{fill}.}
}
\value{
A \code{character} vector of colours or list of \verb{<GridPattern>} objects.
}
\description{
\code{update_alpha()} modifies the transparency of fill colours and/or patterns.
}
\details{
\itemize{
\item Like \code{\link[ggplot2:fill_alpha]{ggplot2::fill_alpha()}} but unlike \code{\link[scales:alpha]{scales::alpha()}} it also attempts
to set the transparency of \verb{<GridPattern>} objects.
\item Unlike \code{\link[ggplot2:fill_alpha]{ggplot2::fill_alpha()}} it will work on a list of length one
with more than one color.
\item \code{update_alpha()} does not depend on \code{ggplot2} or \code{scales}.
}
}
\examples{
# Typical colour input
update_alpha("red", 0.5)

if (utils::packageVersion("grid") > "4.1") {
  # Pattern input
  update_alpha(list(grid::linearGradient()), 0.5)
}
}
