% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patternGrob.R
\name{grid.pattern}
\alias{grid.pattern}
\alias{patternGrob}
\title{Patterned grobs}
\usage{
grid.pattern(
  pattern = "stripe",
  x = c(0, 0.5, 1, 0.5),
  y = c(0.5, 1, 0.5, 0),
  ...,
  id = 1L,
  default.units = "npc",
  prefix = "pattern_",
  name = NULL,
  gp = gpar(),
  draw = TRUE,
  vp = NULL
)

patternGrob(
  pattern = "strip",
  x = c(0, 0.5, 1, 0.5),
  y = c(0.5, 1, 0.5, 0),
  ...,
  id = 1L,
  default.units = "npc",
  prefix = "pattern_",
  name = NULL,
  gp = gpar(),
  draw = TRUE,
  vp = NULL
)
}
\arguments{
\item{pattern}{Name of pattern}

\item{x}{A numeric vector or unit object specifying x-locations.}

\item{y}{A numeric vector or unit object specifying y-locations.}

\item{...}{Pattern parameters}

\item{id}{A numeric vector used to separate locations in \code{x} and
    \code{y} into multiple polygons.  All locations with the same
    \code{id} belong to the same polygon.}

\item{default.units}{A string indicating the default units to use
    if \code{x}, \code{y}, \code{width}, or \code{height}
    are only given as numeric vectors.}

\item{prefix}{Prefix to prepend to the name of each of the pattern parameters in \code{...}}

\item{name}{ A character identifier. }

\item{gp}{An object of class \code{"gpar"}, typically the output
    from a call to the function \code{\link[grid]{gpar}}.  This is basically
    a list of graphical parameter settings.}

\item{draw}{A logical value indicating whether graphics output
    should be produced.}

\item{vp}{A Grid viewport object (or NULL).}
}
\description{
\code{grid.pattern} draws patterned shapes onto the graphic device.
\code{patternGrob} returns the grid grob objects.
}
\examples{
 if (require("grid")) {
   grid.newpage()
   grid.pattern()
   grid.newpage()
   grid.pattern("stripe", colour="blue", fill="yellow", density = 0.5, angle = 135)
   grid.newpage()
   grid.pattern("circle", colour="blue", fill="yellow", size = 2, density = 0.5)
 }
}
\seealso{
\url{https://coolbutuseless.github.io/package/ggpattern/index.html}
         for more details on the patterns and their parameters.
}
