% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pattern-array-ambient.R
\name{grid.pattern_ambient}
\alias{grid.pattern_ambient}
\title{Ambient patterned grobs}
\usage{
grid.pattern_ambient(
  x = c(0.5, 0.067, 0.067, 0.5, 0.933, 0.933),
  y = c(1, 0.75, 0.25, 0, 0.25, 0.75),
  id = 1L,
  ...,
  type = "simplex",
  fill = gp$fill \%||\% "grey80",
  fill2 = "#4169E1",
  frequency = 0.01,
  interpolator = "quintic",
  fractal = "fbm",
  octaves = 3,
  lacunarity = 2,
  gain = 0.5,
  pertubation = "none",
  pertubation_amplitude = 1,
  seed = NA_integer_,
  default.units = "npc",
  name = NULL,
  gp = gpar(),
  draw = TRUE,
  vp = NULL
)
}
\arguments{
\item{x}{A numeric vector or unit object specifying x-locations of the pattern boundary}

\item{y}{A numeric vector or unit object specifying y-locations of the pattern boundary}

\item{id}{A numeric vector used to separate locations in x, y into multiple boundaries.
All locations within the same \code{id} belong to the same boundary.}

\item{...}{Currently ignored}

\item{type}{Either cubic, perlin, simplex, value, white, or worley}

\item{fill}{Fill colour}

\item{fill2}{Second colour}

\item{frequency}{Determines the granularity of the features in the noise.}

\item{interpolator}{How should values between sampled points be calculated?
Either \code{'linear'}, \code{'hermite'}, or \code{'quintic'} (default), ranging from lowest
to highest quality.}

\item{fractal}{The fractal type to use. Either \code{'none'}, \code{'fbm'} (default),
\code{'billow'}, or \code{'rigid-multi'}. It is suggested that you experiment with the
different types to get a feel for how they behaves.}

\item{octaves}{The number of noise layers used to create the fractal noise.
Ignored if \code{fractal = 'none'}. Defaults to \code{3}.}

\item{lacunarity}{The frequency multiplier between successive noise layers
when building fractal noise. Ignored if \code{fractal = 'none'}. Defaults to \code{2}.}

\item{gain}{The relative strength between successive noise layers when
building fractal noise. Ignored if \code{fractal = 'none'}. Defaults to \code{0.5}.}

\item{pertubation}{The pertubation to use. Either \code{'none'} (default),
\code{'normal'}, or \code{'fractal'}. Defines the displacement (warping) of the noise,
with \code{'normal'} giving a smooth warping and \code{'fractal'} giving a more eratic
warping.}

\item{pertubation_amplitude}{The maximal pertubation distance from the
origin. Ignored if \code{pertubation = 'none'}. Defaults to \code{1}.}

\item{seed}{The seed to use for the noise. If \code{NULL} a random seed will be
used}

\item{default.units}{A string indicating the default units to use if \code{x} or \code{y}
are only given as numeric vectors.}

\item{name}{ A character identifier. }

\item{gp}{An object of class \code{"gpar"}, typically the output
    from a call to the function \code{\link[grid]{gpar}}.  This is basically
    a list of graphical parameter settings.}

\item{draw}{A logical value indicating whether graphics output
    should be produced.}

\item{vp}{A Grid viewport object (or NULL).}
}
\value{
A grid grob object invisibly.  If \code{draw} is \code{TRUE} then also draws to the graphic device as a side effect.
}
\description{
\code{grid.pattern_ambient()} draws noise patterns onto the graphic device powered by the \code{ambient} package.
}
\examples{
 if (requireNamespace("ambient")) {
   grid.pattern_ambient(fill = "green", fill2 = "blue")
   grid::grid.newpage()
   grid.pattern_ambient(fill = "green", fill2 = "blue", type = "cubic")
 }
}
\seealso{
For more information about the noise types please see the relevant \code{ambient} documentation:
\code{\link[ambient:noise_cubic]{ambient::noise_cubic()}}, \code{\link[ambient:noise_perlin]{ambient::noise_perlin()}}, \code{\link[ambient:noise_simplex]{ambient::noise_simplex()}},
\code{\link[ambient:noise_value]{ambient::noise_value()}}, \code{\link[ambient:noise_white]{ambient::noise_white()}}, and \code{\link[ambient:noise_worley]{ambient::noise_worley()}}
}
