% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pattern-array-placeholder.R
\docType{data}
\name{grid.pattern_placeholder}
\alias{grid.pattern_placeholder}
\alias{placeholder_names}
\title{Placeholder image patterned grobs}
\format{
An object of class \code{character} of length 26.
}
\usage{
grid.pattern_placeholder(
  x = c(0.5, 0.067, 0.067, 0.5, 0.933, 0.933),
  y = c(1, 0.75, 0.25, 0, 0.25, 0.75),
  id = 1L,
  ...,
  type = "kitten",
  alpha = gp$alpha \%||\% 1,
  aspect_ratio = 1,
  key_scale_factor = 1,
  default.units = "npc",
  name = NULL,
  gp = gpar(),
  draw = TRUE,
  vp = NULL
)

placeholder_names
}
\arguments{
\item{x}{A numeric vector or unit object specifying x-locations of the pattern boundary}

\item{y}{A numeric vector or unit object specifying y-locations of the pattern boundary}

\item{id}{A numeric vector used to separate locations in x, y into multiple boundaries.
All locations within the same \code{id} belong to the same boundary.}

\item{...}{Currently ignored}

\item{type}{Image source.  \code{placeholder_names} is a vector of supported values.
If you would like only greyscale images append \code{bw} to the name.}

\item{alpha}{Alpha (between 0 and 1)}

\item{aspect_ratio}{Override aspect ratio}

\item{key_scale_factor}{Additional scale factor for legend}

\item{default.units}{A string indicating the default units to use if \code{x} or \code{y}
are only given as numeric vectors.}

\item{name}{ A character identifier. }

\item{gp}{An object of class \code{"gpar"}, typically the output
    from a call to the function \code{\link[grid]{gpar}}.  This is basically
    a list of graphical parameter settings.}

\item{draw}{A logical value indicating whether graphics output
    should be produced.}

\item{vp}{A Grid viewport object (or NULL).}
}
\value{
A grid grob object invisibly.  If \code{draw} is \code{TRUE} then also draws to the graphic device as a side effect.
}
\description{
\code{grid.pattern_placeholder()} draws a placeholder image pattern onto the graphic device.
\code{placeholder_names} are character vectors of supported placeholder types.
}
\examples{
  \dontrun{
    # requires internet connection to download from placeholder image websites
    grid.pattern_placeholder(type="bear")
  }
  print(placeholder_names)
}
\seealso{
The \code{ggpattern} documentation: \url{https://coolbutuseless.github.io/package/ggpattern/articles/pattern-placeholder.html}
}
\keyword{datasets}
