% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pattern-geometry-regular_polygon.R
\name{grid.pattern_regular_polygon}
\alias{grid.pattern_regular_polygon}
\title{Regular polygon patterned grobs}
\usage{
grid.pattern_regular_polygon(
  x = c(0, 0, 1, 1),
  y = c(1, 0, 0, 1),
  id = 1L,
  ...,
  colour = gp$col \%||\% "grey20",
  fill = gp$fill \%||\% "grey80",
  angle = 30,
  density = 0.2,
  spacing = 0.05,
  xoffset = 0,
  yoffset = 0,
  scale = 0.5,
  shape = "convex4",
  type = "diagonal",
  rot = 0,
  alpha = gp$alpha \%||\% NA_real_,
  linetype = gp$lty \%||\% 1,
  size = gp$lwd \%||\% 1,
  default.units = "npc",
  name = NULL,
  gp = gpar(),
  draw = TRUE,
  vp = NULL
)
}
\arguments{
\item{x}{A numeric vector or unit object specifying x-locations of the pattern boundary.}

\item{y}{A numeric vector or unit object specifying y-locations of the pattern boundary.}

\item{id}{A numeric vector used to separate locations in x, y into multiple boundaries.
All locations within the same \code{id} belong to the same boundary.}

\item{...}{Currently ignored}

\item{colour}{Stroke colour}

\item{fill}{Fill colour}

\item{angle}{Rotation angle in degrees}

\item{density}{Approx. fraction of area the pattern fills (between 0 and 1)}

\item{spacing}{Spacing between repetitions of pattern (between 0 and 1)}

\item{xoffset}{Shift pattern along x axis (between 0 and 1)}

\item{yoffset}{Shift pattern along y axis (between 0 and 1)}

\item{scale}{For star polygons, multiplier (between 0 and 1)
applied to exterior radius to get interior radius.}

\item{shape}{Either "convex" or "star" followed by the number of exterior vertices
or alternatively "circle", "null", or "square".
For example \code{"convex5"} (default) corresponds to a pentagon
and \code{"star6"} corresponds to a six-pointed star.
The \code{"square"} shape is larger than the \code{"convex4"} shape and is rotated an extra 45 degrees,
it can be used to generate a multi-colored "checkers" effect when density is 1.
The \code{"null"} shape is not drawn, it can be used to create holes within multiple-element patterns.}

\item{type}{Either \code{"diagonal"} (default), \code{"hex"}, or \code{"hex_circle"}.
Adjusts layout, density, and repeating of certain aesthetics in order to aid in achieving
a desired "tiling" or "packing" layout.
\code{"hex"} supports a tiling layout of hexagons (and triangles) while
\code{"hex_circle"} supports a hexagonal packing arrangement for circles.}

\item{rot}{Angle to rotate regular polygon (degrees, counter-clockwise).}

\item{alpha}{Alpha (between 0 and 1) or \code{NA} (default, preserves colors' alpha value)}

\item{linetype}{Stroke linetype}

\item{size}{Stroke linewidth}

\item{default.units}{A string indicating the default units to use if \code{x} or \code{y}
are only given as numeric vectors.}

\item{name}{ A character identifier. }

\item{gp}{An object of class \code{"gpar"}, typically the output
    from a call to the function \code{\link[grid]{gpar}}.  This is basically
    a list of graphical parameter settings.}

\item{draw}{A logical value indicating whether graphics output
    should be produced.}

\item{vp}{A Grid viewport object (or NULL).}
}
\value{
A grid grob object invisibly.  If \code{draw} is \code{TRUE} then also draws to the graphic device as a side effect.
}
\description{
\code{grid.pattern_regular_polygon()} draws a regular polygon pattern onto the graphic device.
}
\examples{
  if (require("grid")) {
    x_hex <- 0.5 + 0.5 * cos(seq(2 * pi / 4, by = 2 * pi / 6, length.out = 6))
    y_hex <- 0.5 + 0.5 * sin(seq(2 * pi / 4, by = 2 * pi / 6, length.out = 6))

    # 'density', 'rot', and 'shape' are vectorized
    grid.newpage()
    grid.pattern_regular_polygon(x_hex, y_hex, colour = "black",
                                 fill = c("blue", "yellow", "red"),
                                 shape = c("convex4", "star8", "circle"),
                                 density = c(0.45, 0.42, 0.4),
                                 spacing = 0.08, angle = 0)

    # checker pattern using "square" shape
    grid.newpage()
    grid.pattern_regular_polygon(x_hex, y_hex, shape = "square",
                                 colour = "transparent",
                                 fill = c("black", "red", "blue", "yellow"),
                                 angle = 0, density = 1.0, spacing = 0.2)

    # checker pattern using the default "convex4" shape
    grid.newpage()
    grid.pattern_regular_polygon(x_hex, y_hex, density = 1.0,
                                 colour = "black", fill = "blue")

    # hexagon tiling
    grid.newpage()
    grid.pattern_regular_polygon(x_hex, y_hex, color = "transparent",
                                 fill = c("white", "grey", "black"),
                                 density = 1.0, spacing = 0.1,
                                 shape = "convex6", type = "hex")

    # triangle tiling
    grid.newpage()
    grid.pattern_regular_polygon(x_hex, y_hex, fill = "green",
                                 density = 1.0, spacing = 0.1,
                                 shape = "convex3", type = "hex")

  }
}
\seealso{
\code{\link[=grid.pattern_circle]{grid.pattern_circle()}} for a special case of this pattern.
The tiling vignette features more examples of regular poylgon tiling using
this function \code{vignette("tiling", package = "gridpattern")}.
}
