% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gridpattern-package.R
\docType{package}
\name{gridpattern-package}
\alias{gridpattern}
\alias{gridpattern-package}
\title{gridpattern: 'grid' Pattern Grobs}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

Provides 'grid' grobs that fill in a user-defined area with various patterns.  Includes enhanced versions of the geometric and image-based patterns originally contained in the 'ggpattern' package as well as original 'pch', 'polygon_tiling', 'regular_polygon', and 'weave' patterns plus support for custom user-defined patterns.
}
\section{Package options}{

The following \code{gridpattern} options may be set globally via \code{\link[base:options]{base::options()}}:
\describe{
\item{ggpattern_geometry_funcs}{Set custom \dQuote{geometry} pattern functions.}
\item{ggpattern_array_funcs}{Set custom \dQuote{array} pattern functions.}
\item{ggpattern_use_R4.1_clipping}{If \code{TRUE} use the grid clipping path feature introduced in R v4.1.0
else do a \code{rasterGrob} approximation of the clipped pattern.
Note not all graphic devices support the grid clipping path feature.}
}
}

\seealso{
Useful links:
\itemize{
  \item \url{https://trevorldavis.com/R/gridpattern/}
  \item \url{https://github.com/trevorld/gridpattern}
  \item Report bugs at \url{https://github.com/trevorld/gridpattern/issues}
}

}
\author{
\strong{Maintainer}: Trevor L Davis \email{trevor.l.davis@gmail.com}

Authors:
\itemize{
  \item Mike FC (Patterns extracted from ggpattern)
}

Other contributors:
\itemize{
  \item Thomas Lin Pedersen (new_data_frame() copied from ggplot2) [contributor]
}

}
\keyword{internal}
