% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pattern_hex.R
\docType{data}
\name{pattern_hex}
\alias{pattern_hex}
\alias{hex_names}
\title{Hex pattern matrix}
\format{
An object of class \code{character} of length 2.
}
\usage{
pattern_hex(type = "hex", subtype = NULL, nrow = 5L, ncol = 5L)

hex_names
}
\arguments{
\item{type}{Either "hex" or "hex_circle".
However for this function there is currently no difference in
behaviour between the two (unlike in \code{\link[=grid.pattern_regular_polygon]{grid.pattern_regular_polygon()}}).}

\item{subtype}{An integer indicating number of colors (or other graphical elements).}

\item{nrow}{Number of rows (height).}

\item{ncol}{Number of columns (width).}
}
\value{
A matrix of integer values indicating where the each color
or other graphical elements should be drawn on a horizontal hex grid
(i.e. hexagons are assumed to be pointy side up).
Indices \verb{[1,1]} of the matrix corresponds to the bottom-left of the grid
while indices \verb{[1,ncol]} corresponds to the bottom-right of the grid.
The even rows are assumed to be on the \strong{left} of the ones on the odd rows
(for those in the same column in the matrix).
This matrix has a "pattern_hex" subclass which supports a special \code{print()} method.
}
\description{
\code{pattern_hex()} returns an integer matrix indicating where each
color (or other graphical element) should be drawn on a (horizontal) hex grid
for a specified hex pattern type and subtype.
\code{hex_names} lists the currently supported hex \code{type}s.
}
\details{
Currently we cycle through elements on the horizontal line and "main" (downward-sloping) diagonal line.
If \code{subtype} is \code{3L} this provides a 3-color hexagonal tiling pattern.
If \code{subtype} is strictly greater than \code{2L} then a hexagon should never touch another hexagon of the same color.
}
\examples{
 # supported hex names
 print(hex_names)

 hex_3color <- pattern_hex("hex", 3L, nrow = 7L, ncol = 9L)
 print(hex_3color)

}
\seealso{
\code{\link[=grid.pattern_regular_polygon]{grid.pattern_regular_polygon()}} for drawing to a graphics device
hexagons, triangles, circles, etc. in hexagon patterns.
}
\keyword{datasets}
